% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_custom.R
\name{cust_ed608_bin}
\alias{cust_ed608_bin}
\title{ED608 Bin Subjects By Challenge Information}
\usage{
cust_ed608_bin(fc, surv, new_groups = c("SUSC", "RES"))
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{surv}{The survival information data frame. Should contain subject id, challenge to infection, and infection status.}

\item{new_groups}{The new group labels you wish to use.}
}
\value{
The Fc Array data frame with the group information replaced with the new group assignments.
}
\description{
Regroup subjects based on whether the survived/remained uninfected for greater or less than the median number of challenges.
}
