% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_custom.R
\name{cust_ed608_pretty_title}
\alias{cust_ed608_pretty_title}
\title{ED608 Convert Feature String To Nicer String}
\usage{
cust_ed608_pretty_title(compound_feat)
}
\arguments{
\item{compound_feat}{A feature string that is ugly and contracted with excess periods, etc.}
}
\value{
The new feature string.
}
\description{
A function that takes in an ugly Fc Array compound title and makes it nicer. Useful if you want axis labels/plot titles that aren't ugly.
}
