% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_custom.R
\name{cust_rv144_bin}
\alias{cust_rv144_bin}
\title{RV144 Bin Subjects By Challenge Information}
\usage{
cust_rv144_bin(fc, samples, surv)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{samples}{The sample mapping informatin data frame. Used for converting sample aliases (4.6.6, etc.) to PIDs, which is what survival information is paired with.}

\item{surv}{The survival information data frame. Should contain subject id, challenge to infection, and infection status.}
}
\value{
The Fc Array data frame with the group information replaced with the new group assignments.
}
\description{
Regroup subjects based on whether the survived/remained uninfected for greater or less than the median number of challenges.
}
