% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_load_config.R
\name{fc_load_config}
\alias{fc_load_config}
\title{Load Fc Analysis Config .R File}
\usage{
fc_load_config(
  config_location,
  experiment_name,
  flags,
  gopts,
  fc = NULL,
  only_group = NULL,
  bin_method = NULL,
  ref_group = NULL
)
}
\arguments{
\item{config_location}{A string specifying the file path to the config.R file, relative to the location from which this function is called, or an absolute file path.}

\item{experiment_name}{A string specifying the name of the current experiment. Will be used to create a separate subdirectory for results.}

\item{fc}{The Fc Array data frame. NULL means that it will be loaded from the location specified in the config.R file. If an Fc Array data frame is passed, then it will be subjected to the filtering, binning, etc. that you have specified.}

\item{bin_method}{If flags$do_bin is TRUE, this is the function you will use to assign new groups.}

\item{ref_group}{If flags$do_differs is TRUE, this is a string specifying the group you want to define as baseline.}
}
\value{
A list containing the Fc Array data frame modified to your specifications, and the new results_dir string for this experiment, etc.
}
\description{
Loads the contents of the provided config file, performs data imputation, any specified filtering, binning, and prints out sessionInfo().
}
