% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_common_utils.R
\name{fccu_extend}
\alias{fccu_extend}
\title{Extend a Vector}
\usage{
fccu_extend(vec, max_len, to_app = NA)
}
\arguments{
\item{vec}{A vector.}

\item{max_len}{How long you want vec to be.}

\item{to_app}{What you want to be appended to vec.}
}
\value{
A vector of length max_len.
}
\description{
Extend vector to match max_len by appending to_app.
}
