% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_common_utils.R
\name{fccu_first_feat_col}
\alias{fccu_first_feat_col}
\title{Find the First Column That Is A Feature}
\usage{
fccu_first_feat_col(fc)
}
\value{
The first column that is not "subject", "group", or "times".
}
\description{
Find the first column that is a feature.
}
