% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_common_utils.R
\name{fccu_match_rows}
\alias{fccu_match_rows}
\title{Match The Length Of A Data Frame And Vector}
\usage{
fccu_match_rows(a, b, data)
}
\arguments{
\item{a}{A vector to be trimmed, must be from data as well (say, rownames()).}

\item{b}{The vector of shorter length that you want length(a) and nrow(data) to be.}

\item{data}{Some data frame you want trimmed.}
}
\value{
A trimmed data frame.
}
\description{
Make data and vector same length/nrow as other vector. Remove entries from data and a until a is the same length as b. By checking if a[i] == b[i] until you find one that isn't.
}
