% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_common_utils.R
\name{fccu_plevels}
\alias{fccu_plevels}
\title{Get Levels for a Non-Factor}
\usage{
fccu_plevels(vec)
}
\arguments{
\item{vec}{The vector you want the "levels" for.}
}
\value{
The result of levels(as.factor(vec)).
}
\description{
This is a shortcut for levels(as.factor(vec)).
}
