% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_data_clean.R
\name{fcdc_impute}
\alias{fcdc_impute}
\title{Missing Value Imputation}
\usage{
fcdc_impute(
  fc,
  start_col = fccu_first_feat_col(fc),
  type = "det",
  method = median
)
}
\arguments{
\item{fc}{The Fc Array data frame. Can be single time point or longitudinal.}

\item{start_col}{Where you want the imputing to start iterating along columns from, presumably fccu_first_feat_col(fc).}

\item{type}{String denoting the type of imputation. Currently, only "det" for deterministic is suppoted.}

\item{method}{For calculating imputed values. Default is median.}
}
\value{
The Fc Array data frame, with missing values replaced based on method. By default, this will be the median of a group's values for that feature.
}
\description{
Impute missing values in an Fc Array data frame on a per-group basis. Thus requires fc$group exists. This function can handle both single time point and longitudinal Fc Array data frames.
}
