% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_data_clean.R
\name{fcdc_normalize}
\alias{fcdc_normalize}
\title{Normalize An Fc Array}
\usage{
fcdc_normalize(fc, center = TRUE, scale = TRUE)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{center}{A boolean or vector. See ?scale}

\item{scale}{A boolean or vector. See ?scale}
}
\value{
The Fc Array data frame, with feature vectors scaled/centered as specified.
}
\description{
Perform normalization on an Fc Array data frame. Just a glorified wrapper for scale() that calls fcdc_impute if necessary.
}
