% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_dim_red.R
\name{fcdr_breadth_score}
\alias{fcdr_breadth_score}
\title{Calculate A Breadth Score}
\usage{
fcdr_breadth_score(fc, to_condense, name = NULL)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{to_condense}{A vector of strings of features that you want to collapse into a single score.}

\item{name}{An optional name parameter to specify for the new score feature. Otherwise, the members of to_condense will be presumed to be of a similar class, and the first substring of the first feature (when split by [.]) will be prepended to "_score".}
}
\value{
The Fc Array data frame, with the features provided replaced with a single score.
}
\description{
Replace a subset of features with a single score representing the geometric mean for each subject across those features. Do NOT pass standardized data.
}
\examples{
# pass only the colnames that contain IgG3
fc_new = fcdr_breath_score(fc, colnames(fc)[grepl("IgG3", colnames(fc), fixed=TRUE)])
# contains the new score feature
fc_new[,"IgG3_score"]

}
