% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_dim_red.R
\name{fcdr_dimred}
\alias{fcdr_dimred}
\title{Perform Dimension Reduction}
\usage{
fcdr_dimred(
  fc,
  method,
  dims = NULL,
  perplexity = 30,
  init_pca = TRUE,
  max_iters = 500,
  theta = 0.5,
  seed = 1337,
  verbose = FALSE
)
}
\arguments{
\item{fc}{The Fc Array data frame}

\item{method}{A string "pca", "tsne", or "umap" the denotes what dimension reduction method will be used.}

\item{dims}{The number of dimensions to reduce the dataset down to. Pass NULL if you want all components back (PCA), or if you feel like letting tSNE and UMAP yell at you that they picked 2 for you.}

\item{perplexity}{A parameter for tSNE. It will be adjusted for you if you don't provide it/change it and it needs providing/to be changed.}

\item{init_pca}{A boolean that determines whether or not tSNE performs an initial pca to perform a first-pass on reducing dimensions.}

\item{max_iters}{The maximum number of iterations you want tSNE or UMAP to perform before terminating.}

\item{theta}{Speed/accuracy tradeoff for tSNE. 0.0 is exact tSNE, 1.0 is "just get it done ASAP idfc if it's wildly inaccurate".}

\item{seed}{The random seed you wish to use to facilitate reproducibility.}

\item{verbose}{A boolean that determines how much you get yelled at by tSNE and UMAP about how they're doing.}
}
\value{
The Fc Array data frame with all the features replaced by all the components generated by the method you used.
}
\description{
Function for doing multiple kinds of dimension reduction.
}
\examples{
colnames(fc)
# [1] "subject" "group" "IgG3.gp41" "IgG.gp70.V1.V2" # etc.
fc_tsne = fcdr_dimred(fc, "tsne", dims=2)
colnames(fc_tsne)
# [1] "subject" "group" "tsne.1" "tsne.2"

}
