% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_dim_red.R
\name{fcdr_fc_breadth}
\alias{fcdr_fc_breadth}
\title{Calculate Many Breadth Scores}
\usage{
fcdr_fc_breadth(fc, to_condense)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{to_condense}{A vector of strings of features that you want to collapse into individual scores.}
}
\value{
The Fc Array data frame, with each class of features provided replaced with individual scores. None of the other feature data remains.
}
\description{
Replace each subset of features with a single score representing the geometric mean for each subject across those features. Do NOT pass standardized data. Does not work with ordering subclasses (IgG1 before IgG2, etc. - see how fcgo_get_prop works).
}
\examples{
# pass only the colnames that contain IgG3
gopts$feat_ant = c("gp120","gp140","V1.V2.","gag","p17,p24,p51,p55,p66")
# contains gp120_score, gp140_score, V1.V2._score, etc.
fc_new = fcdr_breath_score(fc, gopts$feat_ant)

}
