% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_feat_filt.R
\name{fcff_exclude_attr}
\alias{fcff_exclude_attr}
\title{Exclude An Attribute}
\usage{
fcff_exclude_attr(fc, attr, clm)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{attr}{The value of the thing you want gone.}

\item{clm}{What column in fc attr will be found in.}
}
\value{
The Fc Array data frame, without the rows that have values in clm matching attr.
}
\description{
Exclude an attribute from an Fc Array data frame.
}
\examples{
fc_no_plac = fcff_exclude_attr(fc, "PLACEBO", "group")

}
