% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_feat_filt.R
\name{fcff_feats_diff}
\alias{fcff_feats_diff}
\title{Filter Features By Difference From Baseline}
\usage{
fcff_feats_diff(fc, diff_opts)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{diff_opts}{A list containing the following sub-parameters:
\itemize{
    \item diff_opts$base_ids Uhe row indices in fc that correspond to the subjects who represent baseline (say, those in the PLACEBO group).
    \item diff_opts$test Which statistical test to use to determine difference from baseline.
    \item diff_opts$alternative Alternative hypothesis (two.sided, either (two.sided/2)).
    \item diff_opts$adj_method The method you want to use to adjust p-values. Use "none" as a passthrough.
}}
}
\value{
The p-values that represent the comparison between baseline and the other subjects using the test specified in diff_opts.
}
\description{
Filter features from an Fc Array data frame by their difference from baseline.
}
\examples{
p_vals = fcff_feats_diff(fc, diff_opts)
keep = p_vals < 0.05

}
