% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_feat_filt.R
\name{fcff_filter_features}
\alias{fcff_filter_features}
\title{Filter Features}
\usage{
fcff_filter_features(
  fc,
  feats,
  action,
  behavior = "permissive",
  diff_opts = NULL
)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{feats}{Vector of strings for the features you want to keep or remove.}

\item{action}{String specifying whether you want to "keep" or "toss" the features in feats.}

\item{behavior}{Whether the feature names will be matched by:
\itemize{
    \item containing any of the strings in feats ("permissive").
\item containing all of the strings in feats ("strict").
\item Alternatively, "differs" specifies that features will be removed by difference from baseline. This will be done instead of keyword filtering (so to do both, call this function twice).
}}

\item{diff_opts}{list of options if behavior = "differs". see ?fcff_feats_diff.}
}
\value{
The Fc Array data frame filtered based on what you specified.
}
\description{
Filter features based on keyword. You can keep those, remove those, and many other things.
}
\examples{
new_fc = fcff_filter_features(fc, c("gp41", "gp120"), "keep")
my_diff_opts = list() # fill with stuff
diff_fc = fccu_filter_features(fc, NULL, NULL, behavior="differs",
					diff_opts=my_diff_opts)

}
