% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_feat_filt.R
\name{fcff_only_attr}
\alias{fcff_only_attr}
\title{Keep Only Specific Values Of A Column}
\usage{
fcff_only_attr(fc, attr, attrs, clm)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{attr}{The value of the thing you want gone.}

\item{attrs}{A vector containing all the values that attr can take on in clm.}

\item{clm}{What column in fc attr will be found in.}
}
\value{
The Fc Array data frame, with only the rows that have values in clm matching attr.
}
\description{
A wrapper for fcff_exclude_attr that performs the exclusion behavior repeatedly in order to produce a dataset with just attr in attrs.
}
\examples{
groups = c("PLACEBO", "VACCINE")
fc_only_placebo = fcff_only_attr(fc, "PLACEBO", groups, "group")

}
