% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_feat_filt.R
\name{fcff_remove_cor}
\alias{fcff_remove_cor}
\title{Remove Correlated Features}
\usage{
fcff_remove_cor(fc, cutoff, behavior = "sample")
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{cutoff}{The correlation coefficient cutoff above which we will remove features from fc.}

\item{behavior}{How we'll decide what to keep as a representative feature from those to be removed. "sample" randomly selects one feature. "first" picks the first one in the list.}
}
\value{
The Fc Array data frame filtered based on what you specified.
}
\description{
Remove correlated features from fc based on cutoff.
}
