% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_graphic_output.R
\name{fcgo_boxplot}
\alias{fcgo_boxplot}
\title{Generate Boxplots}
\usage{
fcgo_boxplot(fc, to_box, results_dir, flags, gopts, seed = 1337)
}
\arguments{
\item{fc}{The Fc Array data frame}

\item{to_box}{A vector of strings of feature names that you wish to generate boxplots for.}

\item{results_dir}{Where you want your results to be saved.}

\item{flags}{A list of analysis booleans. Must have specified the subparameters:
\itemize{
\item flags$do_same_range Do you want all boxplots to have the same y-limits?
\item flags$nicer_legend Do you want a nicer legend? Currently applies only to the hardcoded case of RV144, where the binning generates groups that are "yes" and "no". If this flag is TRUE, these get replaced in the plot legends with "Infected" and "Uninfected".
}}

\item{gopts}{A list of graphical options. Must have specified the subparameters, and then the subsequent optional parameters:
\itemize{
\item gopts$bpar_mar The margins for the plots. Sensible starting place is c(5.1, 4.1, 4.1, 4).
\item gopts$jplot_group_cols The group colors for the jitter plot points that get drawn on top of the boxes. Should be the same length as unique(fc$group) (i.e. the number of groups you have).
\item gopts$bplot_group_cols The same as jplot_group_cols, but for the boxes themselves. Will need to be slightly different if you want points to show up. If one of jplot_group_cols is black, for example, the corresponding bplot_group_cols should be white. Experiment to get it right.
\item gopts$box_legend_inset The inset for the legend in the boxplots. Sensible starting place is c(-0.40, 0).
}
\itemize{
\item gopts$pdf_name_suffix If specified, will change the output pdf name from "group1_group2_boxplots.pdf" to "group1_group2_[pdf_name_suffix]_boxplots.pdf".
\item gopts$main If specified, sets the title of all the boxplots to the provided string. Else, the name of the feature is used.
\item gopts$ylab If specified, sets the title of all boxplots to the provided string. Else, "MFI" will be used.
}}
}
\value{
Nothing! Just outputs plots.
}
\description{
Generate boxplots for a subset of features in an Fc Array data frame, separated out by group membership. Plots both boxes and whiskers and jitter plot of the points. Supports more than 2 groups. Resultant plots will live in a pdf, and be saved in results_dir.
}
\examples{
flags = list() # add some things
gopts = list() # add some more things
# produce a pdf containing boxplots of all IgG3 features, with points colored by fc$group membership
fcgo_boxplot(fc, colnames(fc)[grepl("IgG3", colnames(fc), fixed=TRUE)], "/home/me/results/", flags, gopts)

}
