% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_graphic_output.R
\name{fcgo_l10_pplot}
\alias{fcgo_l10_pplot}
\title{Plot -log10 Transformed p-values}
\usage{
fcgo_l10_pplot(
  p_vals,
  pdf_name,
  main,
  line_method,
  gopts,
  pt_colors,
  pt_shapes,
  ylim = NULL
)
}
\arguments{
\item{p_vals}{A vector of p-values.}

\item{pdf_name}{The name of the output file where the plot will be saved. Should be a full file path.}

\item{main}{The title of the plot.}

\item{line_method}{A string indicating the method you want to use to draw the arbitrary cutoff line. "bonferroni" (for p < 0.05), "fdr" (for q < 0.2), and "raw" (just -log10(0.05)) are supported.}

\item{gopts}{A list of graphical options. Must have specified the subparameters:
\itemize{
\item gopts$ppar_mar The margins for this plot. See ?par for more info. Default is c(5.1, 4.1, 4.1, 6).
\item gopts$reag_legend_inset The inset for the legend relative to the plot.
\item gopts$reag_cats The categories for your detection reagants.
\item gopts$reag_cols The colors for the detection reagant categories.
\item gopts$ant_shapes The shapes for the antigens being detected.
}}

\item{pt_colors}{Colors for each point in p_vals. These should be generated from gopts$reag_cols. See ?fcgo_get_prop for an example of how to generate this.}

\item{pt_shapes}{Shapes for each point in p_vals. These should be generated from gopts$ant_shapes. See ?fcgo_get_prop for an example of how to generate this.}

\item{ylim}{An optional parameter for setting the y-limits of the plot.}
}
\value{
Nothing!
}
\description{
Plots p-values after -log10 transforming them. This makes a plot where points higher up on the y-axis are lower p-values. Points are colored by detection reagant and shaped by antigen being detected.
}
