% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_graphic_output.R
\name{fcgo_volcano_plot}
\alias{fcgo_volcano_plot}
\title{Make A Volcano Plot}
\usage{
fcgo_volcano_plot(
  foch,
  p_vals,
  line_method,
  main,
  xlab,
  ylab,
  gopts,
  pt_colors,
  pt_shapes
)
}
\arguments{
\item{foch}{A vector of fold change values.}

\item{p_vals}{A vector of p-values.}

\item{line_method}{A string indicating the method you want to use to draw the arbitrary cutoff line. "bonferroni" (for p < 0.05), "fdr" (for q < 0.2), and "raw" (just -log10(0.05)) are supported.}

\item{main}{The title of the plot.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{gopts}{A list of graphical options. Must have specified the subparameters:
\itemize{
\item gopts$volc_xlim The x-limits for the volcano plot.
\item gopts$volc_ylim The y-limits for the volcano plot.
\item gopts$reag_legend_inset The inset for the legend relative to the plot. Reasonable place to start is c(-0.18, 0).
\item gopts$reag_cats The categories for your detection reagants.
\item gopts$reag_cols The colors for the detection reagant categories.
\item gopts$ant_shapes The shapes for the antigens being detected.
}}

\item{pt_colors}{Colors for each point in p_vals. These should be generated from gopts$reag_cols. See ?fcgo_get_prop for an example of how to generate this.}

\item{pt_shapes}{Shapes for each point in p_vals. These should be generated from gopts$ant_shapes. See ?fcgo_get_prop for an example of how to generate this.}
}
\value{
Nothing!
}
\description{
Plots p-values versus fold change values. Points are colored by detection reagant and shaped by antigen being detected.
}
