% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_assert_input}
\alias{fch_assert_input}
\title{Check The Validity Of Fc Array}
\usage{
fch_assert_input(fc)
}
\arguments{
\item{fc}{The Fc Array data frame.}
}
\value{
A boolean determining whether the Fc Array passed inspection.
}
\description{
Confirm that an Fc Array data frame has only 2 groups and only one time point.
}
