% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_covb_simple}
\alias{fch_covb_simple}
\title{Calculate Simple Change Over Baseline}
\usage{
fch_covb_simple(base_measure, measure)
}
\arguments{
\item{base_measure}{The value for this data point at baseline}

\item{measure}{The value for this data point at some later time point}
}
\value{
A single value: measure/base_measure
}
\description{
Calculate Simple Change Over Baseline
}
