% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_fc_comb}
\alias{fch_fc_comb}
\title{Generate A Single Fc Array Data Frame Of Just Two Groups}
\usage{
fch_fc_comb(fc, group_combs, to_comp)
}
\arguments{
\item{fc}{The Fc Array data frame. Must be a single time point.}

\item{group_combs}{The group combination matrix obtained from combn(groups, 2).}

\item{to_comp}{The column in group_combs that is to be kept.}
}
\value{
The Fc Array data frame with just the two groups in to_comp.
}
\description{
Generate A Single Fc Array Data Frame Of Just Two Groups
}
