% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_fcs_assert_input}
\alias{fch_fcs_assert_input}
\title{Check The Validity Of Many Fc Arrays}
\usage{
fch_fcs_assert_input(fcs)
}
\arguments{
\item{fc}{The Fc Array data frame.}
}
\value{
A boolean determining whether the Fc Array passed inspection.
}
\description{
Confirm that each Fc Array in the list has only 2 groups, only one time point, and the list has been given names.
}
