% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_geom_mean}
\alias{fch_geom_mean}
\title{Caluclate Geometric Mean}
\usage{
fch_geom_mean(vec)
}
\arguments{
\item{vec}{A vector of numeric values.}
}
\value{
A single value: the nth root of the product of vec
}
\description{
Caluclate Geometric Mean
}
