% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_impute_fc_time}
\alias{fch_impute_fc_time}
\title{Missing Value Imputation For A Single Time Point}
\usage{
fch_impute_fc_time(fc, start_col, method = median)
}
\arguments{
\item{fc}{The Fc Array data frame, but only at a single time point.}

\item{start_col}{Where you want the imputing to start iterating along columns from, presumably fccu_first_feat_col(fc).}

\item{method}{For calculating imputed values. Default is median.}
}
\value{
The Fc Array data frame, with missing values replaced based on method. By default, this will be the median of a group's values for that feature.
}
\description{
Missing Value Imputation For A Single Time Point
}
