% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_make_reag_legend}
\alias{fch_make_reag_legend}
\title{Make A Reagant/Antigen Plot Legend}
\usage{
fch_make_reag_legend(gopts)
}
\arguments{
\item{gopts}{A list of graphical options. Must have specified the subparameters:
\itemize{
\item gopts$reag_legend_inset The inset for the legend relative to the plot.
\item gopts$reag_cats The categories for your detection reagants.
\item gopts$reag_cols The colors for the detection reagant categories.
\item gopts$ant_shapes The shapes for the antigens being detected.
}}
}
\value{
Nothing!
}
\description{
Add a reagant/antigen legend to a plot.
}
