% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_min_max_feat}
\alias{fch_min_max_feat}
\title{Calculate Many Min-Max Z-Scores}
\usage{
fch_min_max_feat(vec)
}
\arguments{
\item{vec}{A vector you wish to min-max z-scale.}
}
\value{
A vector of min-max z-scores.
}
\description{
Calculate Many Min-Max Z-Scores
}
