% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_helpers.R
\name{fch_z_score}
\alias{fch_z_score}
\title{Calculate A Single Min-Max Z-Score}
\usage{
fch_z_score(x, vec)
}
\arguments{
\item{x}{A single data point.}

\item{vec}{The vector from which x derives.}
}
\value{
A single min-max z-score for x.
}
\description{
Calculate A Single Min-Max Z-Score
}
