% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_multi.R
\name{fcmu_fcs_combs}
\alias{fcmu_fcs_combs}
\title{Generate Groupwise Combinations Of Fc Arrays}
\usage{
fcmu_fcs_combs(fc, order, num_groups)
}
\arguments{
\item{fc}{The Fc Array data frame.}

\item{order}{Determines what order the data frames are in the list
\itemize{
\item "groups" (g1v2a, g1v2b, g1v3a, g1v3b, g2v3a, g2v3b, etc...).
\item "times" (g1v2a, g1v3a, g2v3a, g1v2b, g1v3b, g2v3b, etc...).
}}

\item{num_groups}{The number of groups to represent in each entry in the returned list. For example, 2 would mean all pairwise combinations of groups would be represented.}
}
\value{
A list of Fc Array data frames, each with only num_groups represented. Returns NULL if it encountered an error.
}
\description{
Generate a list of data frames that each consist of only two treatment groups. Makes many group comparisons easier when paired with for loops.
}
