% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_multi.R
\name{fcmu_fcs_find_ylim}
\alias{fcmu_fcs_find_ylim}
\title{Find y-limits For Many Data Frames}
\usage{
fcmu_fcs_find_ylim(data_list, clm)
}
\arguments{
\item{data_list}{A list of data frames.}

\item{clm}{Which column in each data frame you are interested in.}
}
\value{
A vector of y-limits c(lower, upper). If, for some reason, no y-limits were found, you get back NULL and an error is written. If only a y-maximum was found, the y-minimum is set to 0 and you get back c(0, y_max).
}
\description{
Find the y-limits that will encompass all of the data present in the data frames in the list. Allows plots to be drawn with the same y-limits, making visual inspection and comparison easier and less dubious.
}
