% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_multi.R
\name{fcmu_fcs_fochs}
\alias{fcmu_fcs_fochs}
\title{Perform fcst_fochs On Many Fc Array Datasets}
\usage{
fcmu_fcs_fochs(fcs, method = median)
}
\arguments{
\item{fcs}{A list of Fc Array data frames. Each data frame should have just one time point and just two groups. The names of the entries in fcs should be a contraction of the groups you are comparing, separated by an underscore. The first group in this string will be the numerator in the fold change ratios.}

\item{method}{What method will be used to calculate fold change. The default is the median of each group's values.}
}
\value{
A list containing data frames for each corresponding Fc Array data frame in fcs. Each data frame has fold change values for each feature (features are columns).
}
\description{
For each feature in each Fc Array data frame, calculate the fold change between the measurements of one group and another using method. Which group is the numerator is determined by the structure of comp_name.
}
