% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_multi.R
\name{fcmu_fcs_mannwhit}
\alias{fcmu_fcs_mannwhit}
\title{Perform fcst_mannwhit On Many Fc Array Datasets}
\usage{
fcmu_fcs_mannwhit(fcs, adj_method, alternative = "two.sided")
}
\arguments{
\item{fcs}{A list of Fc Array data frames. Each data frame should have just one time point and just two groups. The names of the entries in fcs should be a contraction of the groups you are comparing, separated by an underscore.}

\item{adj_method}{The method you want to use to adjust p-values. Use "none" as a passthrough.}

\item{alternative}{What alternative hypothesis we're testing. Supports "two.sided", "one.sided", and "either" (result of "two.sided"/2).}

\item{results_dir}{A string representing the directory where you want to store the results of your analyses.}
}
\value{
A list containing data frames for each corresponding Fc Array data frame in fcs. Each data frame has p-values for each feature (features are columns).
}
\description{
For each feature in each Fc Array data frame, perform the Mann-Whitney U (Wilcoxon Rank-Sum) test to test the null hypothesis that a randomly selected value from one population (group) will be less than or greater than a randomly selected value from a second population (group). Here, this is to test whether the two sets of samples, which should be independent (treatment group A should not have an impact on treatment group B), were selected from populations having the same distribution, and there thus being no difference in the groups.
}
