% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_stat_model.R
\name{fcsm_test_model}
\alias{fcsm_test_model}
\title{Test A Statistical Model}
\usage{
fcsm_test_model(model, test, test_group)
}
\arguments{
\item{model}{The model you generated, be it from lm, glm, etc.}

\item{test}{The test data frame you'll be using. Should not include any group membership information.}

\item{test_group}{The actual group membership for your test data.}
}
\value{
The accuracy of your model when predicting on the test data.
}
\description{
A wrapper for testing statistical models.
}
