% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_stat_model.R
\name{fcsm_uni}
\alias{fcsm_uni}
\title{Perform Univariate Analysis}
\usage{
fcsm_uni(fc, adj_method)
}
\arguments{
\item{fc}{The Fc Array data frame. Will be standardized for you so you are comparing apples to apples.}

\item{adj_method}{The method you want to use to adjust p-values. Use "none" as a passthrough.}
}
\description{
Perform univariate analysis on each feature in the Fc Array data frame. Return both p-values and q-values (based on adj_method).
}
