% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_stat_test.R
\name{fcst_covb}
\alias{fcst_covb}
\title{Calculate Change Over Baseline}
\usage{
fcst_covb(fc, method = fch_covb_simple)
}
\arguments{
\item{fc}{The Fc Array data frame, in longitudinal format.}

\item{method}{The method used to calculate change over baseline. The default is a simple division.}
}
\value{
The Fc Array data frame, in longitudinal format (sans the first time point, which is baseline), with all feature MFI values replaced by change over baseline values as calculated by method.
}
\description{
Calculate change over baseline for each subject using a provided method.
}
