% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_stat_test.R
\name{fcst_mannwhit}
\alias{fcst_mannwhit}
\title{Perform Mann-Whitney U Test On Fc Array Features}
\usage{
fcst_mannwhit(fc, adj_method, alternative = "two.sided", exact = FALSE)
}
\arguments{
\item{fc}{The Fc Array data frame. Must have two groups present.}

\item{adj_method}{The method you want to use to adjust p-values. Use "none" as a passthrough.}

\item{alternative}{What alternative hypothesis we're testing. Supports "two.sided", "one.sided", and "either" (result of "two.sided"/2).}

\item{exact}{A boolean to control whether or not exact p-values are calculated.}
}
\value{
A vector of p-values of the same length as fccu_first_feat_col(fc):ncol(fc).
}
\description{
For each feature in an Fc Array data frame, perform the Mann-Whitney U (Wilcoxon Rank-Sum) test to test the null hypothesis that a randomly selected value from one population (group) will be less than or greater than a randomly selected value from a second population (group). Here, this is to test whether the two sets of samples, which should be independent (treatment group A should not have an impact on treatment group B), were selected from populations having the same distribution, and there thus being no difference in the groups.
}
